LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/abspos-containing-block-outside-scroller-expected.txt


FAIL Abs-pos descendant with containing block outside the scroller. assert_equals: expected 50 but got 25
